refact: update&delete friend move to friendship service, controller
